+++ Dragon.Serverless.API.IntegrationTests/Books/GetBook.cs
@@ -11,8 +11,8 @@ namespace Dragon.Serverless.API.IntegrationTests.Books
         private readonly string baseUrl = TestContext.Parameters["ApiBaseUrl"];
         private readonly string shopName = TestContext.Parameters["ShopName"];
 
-        private const int EXISTING_BOOK = 100;
-        private const int UNEXISTING_BOOK = 99999;
+        private const long EXISTING_BOOK = 100;
+        private const long UNEXISTING_BOOK = 99999;
 
         private RestClient restClient;
