+++ Dragon.Serverless.API.IntegrationTests/Sales/UnsellBook.cs
@@ -13,10 +13,10 @@ namespace Dragon.Serverless.API.IntegrationTests.Sales
         private readonly string shopName = TestContext.Parameters["ShopName"];
         private readonly string baseUrl = TestContext.Parameters["ApiBaseUrl"];
 
-        private const int EXISTING_BOOK = 100;
-        private const int EXISTING_BOOK2 = 101;
-        private const int NON_EXISTING_BOOK = 99999;
-        private const int FULLY_AVAILABLE_BOOK = 160;
+        private const long EXISTING_BOOK = 100;
+        private const long EXISTING_BOOK2 = 101;
+        private const long NON_EXISTING_BOOK = 99999;
+        private const long FULLY_AVAILABLE_BOOK = 160;
 
         private RestClient restClient;
 
@@ -75,7 +75,7 @@ namespace Dragon.Serverless.API.IntegrationTests.Sales
             Assert.AreEqual(HttpStatusCode.BadRequest, response.StatusCode);
         }
 
-        private string GetLastSaleDate(string shop, int isbn)
+        private string GetLastSaleDate(string shop, long isbn)
         {
             var request = SaleRequestHelper.GetBookSales(shop, isbn);
             var response = this.restClient.Execute<List<SaleResponse>>(request);
