+++ Dragon.Serverless.API/Functions/Sales/SellBook.cs
@@ -22,7 +22,7 @@ namespace Dragon.Serverless.API.Functions.Sales
         public async Task<IActionResult> Run(
             [HttpTrigger(AuthorizationLevel.Function, "post", Route = null)] HttpRequest req)
         {
-            if (!int.TryParse(req.Query["isbn"], out int isbn) || string.IsNullOrWhiteSpace(req.Query["shop"]))
+            if (!long.TryParse(req.Query["isbn"], out long isbn) || string.IsNullOrWhiteSpace(req.Query["shop"]))
                 return new BadRequestResult();
 
             var result = await this.saleAppService.CreateAsync(req.Query["shop"], isbn, req.Query["seller"]);
