+++ Dragon.Serverless.API/Functions/Sales/UnsellBook.cs
@@ -22,7 +22,7 @@ namespace Dragon.Serverless.API.Functions.Sales
         public async Task<IActionResult> Run(
             [HttpTrigger(AuthorizationLevel.Function, "delete", Route = null)] HttpRequest req)
         {
-            if (!int.TryParse(req.Query["isbn"], out int isbn)
+            if (!long.TryParse(req.Query["isbn"], out long isbn)
                 || string.IsNullOrWhiteSpace(req.Query["date"])
                 || string.IsNullOrWhiteSpace(req.Query["shop"]))
                 return new BadRequestResult();
